% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterStats.R
\name{clusterStats}
\alias{clusterStats}
\title{Compute LVR and meandiff statistics for beta values after batch
correction}
\usage{
clusterStats(pre_betas, post_betas, kClusters)
}
\arguments{
\item{pre_betas}{a matrix of methylation beta values \bold{prior to}
correction.}

\item{post_betas}{a matrix of methylation beta values \bold{after}
correction.}

\item{kClusters}{a kClusters S3 object}
}
\value{
A \code{data.frame} containing clustering stats.
}
\description{
Illumina Infinium BeadChip
}
\details{
Betas values should be of type \code{double} with samples in
columns and betas in rows. The betas need to be bounded between 0 and 1.
The matrix is typically exported from a \code{\link{GenomicRatioSet}},
\code{\link{GenomicMethylSet}} or \code{\link{MethylSet}} object via the
\code{getBeta} S4 accessor method.
}
\examples{
library(Harman)
library(HarmanData)
data(episcope)
bad_batches <- c(1, 5, 9, 17, 25)
is_bad_sample <- episcope$pd$array_num \%in\% bad_batches
myK <- discoverClusteredMethylation(episcope$original[, !is_bad_sample])
mykClust = kClusterMethylation(episcope$original, row_ks=myK)
res = clusterStats(pre_betas=episcope$original,
                   post_betas=episcope$harman,
                   kClusters = mykClust)
all.equal(episcope$ref_md$meandiffs_harman, res$meandiffs)
all.equal(episcope$ref_lvr$var_ratio_harman, res$log2_var_ratio)
}
\seealso{
\code{\link{kClusterMethylation}},
\code{\link{discoverClusteredMethylation}}
}
