% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detachHarman.R
\name{detachHarman}
\alias{detachHarman}
\title{Detach the Harman package and its shared C/C++ library code}
\usage{
detachHarman()
}
\value{
None
}
\description{
A helper function that can be called if \code{\link{harman}}
had to be aborted.
}

