% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discoverClusteredMethylation.R
\name{discoverClusteredMethylation}
\alias{discoverClusteredMethylation}
\title{Discover clustered beta values}
\usage{
discoverClusteredMethylation(
  betas,
  ks = 1:10,
  min_cluster_size = 5,
  min_cluster_dist = 0.1,
  max_clusters = 4,
  cores = 1,
  printInfo = FALSE
)
}
\arguments{
\item{betas}{a matrix of methylation beta values with samples in columns and
betas in rows.}

\item{ks}{integer, the range of k's to consider for clustering (defaults to
1:10).}

\item{min_cluster_size}{integer, the minimum number of samples in a cluster
(defaults to 5).}

\item{min_cluster_dist}{numeric, the minimum beta difference in cluster
centroids (defaults to 0.1).}

\item{max_clusters}{integer, the maximum value of k returned (defaults to 4).}

\item{cores}{integer, specifies the number of cores to use for computation.}

\item{printInfo}{logical, whether to print information during computation or
not.}
}
\value{
A named vector containing the optimal value for k.
}
\description{
This function takes a matrix of methylation beta values
(typically from the Illumina Infinium Methylation Assay) and clusters the
data across a range of ks specified buy the user.

Then the data is reclustered again across the the best two candidate values
for k (determined by the rate of change in Bayesian information criterion),
and minimum cluster size and distance filters are employed. If both clusters
meet these filters, then the higher value of k is returned.  
This function should be run on uncorrected data that ideally has slides
removed which are prone to batch effect. This will bias towards finding
clusters that are driven by biological factors such as X-chromosome
inactivation and allele-specific methylation.

The output of this function is input for the
\code{\link{kClusterMethylation}} function which extracts cluster membership
and statistics on variance for a given matrix of beta values. It might be
useful to discover clusters on samples less prone to clustering due to batch
effect or cellular heterogeneity and then recluster all the data for set
values of k via the \code{\link{kClusterMethylation}} function.

Finally, a comparison of differences of uncorrected to
batch-corrected beta values can be made using \code{\link{clusterStats}}.
This function generates a data.frame containing log variance ratio and mean
beta differences to clusters after correction.
}
\details{
Betas values should be of type \code{double} with samples in
columns and betas in rows. The betas need to be bounded between 0 and 1.
The matrix is typically exported from a \code{\link{GenomicRatioSet}},
\code{\link{GenomicMethylSet}} or \code{\link{MethylSet}} object via the
\code{getBeta} S4 accessor method.
}
\examples{
library(Harman)
library(HarmanData)
data(episcope)
bad_batches <- c(1, 5, 9, 17, 25)
is_bad_sample <- episcope$pd$array_num \%in\% bad_batches
myK <- discoverClusteredMethylation(episcope$original[, !is_bad_sample])
mykClust = kClusterMethylation(episcope$original, row_ks=myK)
res = clusterStats(pre_betas=episcope$original,
                   post_betas=episcope$harman,
                   kClusters = mykClust)
all.equal(episcope$ref_md$meandiffs_harman, res$meandiffs)
all.equal(episcope$ref_lvr$var_ratio_harman, res$log2_var_ratio)
}
\seealso{
\code{\link{kClusterMethylation}}, \code{\link{clusterStats}}
}
