% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmanScores.R
\name{harmanScores}
\alias{harmanScores}
\title{A Principal components prcomp function tweaked for Harman}
\usage{
harmanScores(x)
}
\arguments{
\item{x}{matrix, data matrix of values to perform PCA on.}
}
\value{
scores, a prcomp-like object with rotation, scores and the center
values. The scores are corrected, but all three are needed later to
reconstruct the data.
}
\description{
A tweaking of \code{stats::prcomp} such that for the svd, the
transpose of u is used instead of v when the number of assays is less than
the number of samples.
}

