% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.harmanresults.R
\name{plot.harmanresults}
\alias{plot.harmanresults}
\title{Plot method for harman}
\usage{
\method{plot}{harmanresults}(x, ...)
}
\arguments{
\item{x}{An instance of \code{harmanresults}.}

\item{...}{further plotting parameters.}
}
\value{
None
}
\description{
Plot method for instances of \code{\link{harmanresults}}.
}
\seealso{
\code{\link{harmanresults}} \code{\link{pcaPlot}}
}

