% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.harmanresults.R
\name{print.summary.harmanresults}
\alias{print.summary.harmanresults}
\title{Printing Harmanresults summaries.}
\usage{
\method{print}{summary.harmanresults}(x, ...)
}
\arguments{
\item{x}{an object of class \code{summary.harmanresults}, usually, a result
of a call to \code{summary.harmanresults}.}

\item{...}{further parameters.}
}
\value{
Prints summary information from an object of class
\code{summary.harmanresults}.
}
\description{
Print method for \code{summary.harmanresults}.
}

