% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftBetas.R
\name{shiftBetas}
\alias{shiftBetas}
\title{Shift beta values from 0 and 1 to avoid infinite M values}
\usage{
shiftBetas(betas, shiftBy = 1e-04)
}
\arguments{
\item{betas}{matrix, beta values.}

\item{shiftBy}{numeric, the amount to shift values of \code{0} and
\code{1} by.}
}
\value{
None
}
\description{
A convienance function for methylation data.
}
\examples{
betas <- seq(0, 1, by=0.05)
range(betas)
newBetas <- shiftBetas(betas, shiftBy=1e-4)
newBetas
range(newBetas)
}

