% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.harmanresults.R
\name{summary.harmanresults}
\alias{summary.harmanresults}
\title{Summarizing harman results.}
\usage{
\method{summary}{harmanresults}(object, ...)
}
\arguments{
\item{object}{An object of class \code{harmanresults}.}

\item{...}{further parameters.}
}
\value{
Returns an object of class \code{summary.harmanresults}.
}
\description{
Summary method for class \code{\link{harmanresults}}.
}
\examples{
library(HarmanData)
data(OLF)
expt <- olf.info$Treatment
batch <- olf.info$Batch
olf.harman <- harman(olf.data, expt, batch)
summary(olf.harman)
}
\seealso{
\code{\link{harmanresults}}
}

